refactor(config): split config.py into config/ package with shared mixins
Split 414-LOC module into:
- config/base.py: CopilotSettingsMixin, PromptSettingsMixin, DispatchSettingsMixin
- config/settings.py: Settings, JiraSettings (controller config)
- config/runner_settings.py: TaskRunnerSettings (worker config)
- config/validators.py: shared parse_comma_list validator
- config/__init__.py: re-exports Settings, TaskRunnerSettings, JiraSettings

The three mixins eliminate ~22 duplicated field definitions between
Settings and TaskRunnerSettings. Both classes now compose from the
same field groups via Pydantic v2 multiple inheritance.
