+++ src/gitlab_copilot_agent/config/__init__.py
@@ -1,0 +1,10 @@
+"""Application configuration — Settings, TaskRunnerSettings, JiraSettings."""
+
+from gitlab_copilot_agent.config.runner_settings import TaskRunnerSettings
+from gitlab_copilot_agent.config.settings import JiraSettings, Settings
+
+__all__ = [
+    "JiraSettings",
+    "Settings",
+    "TaskRunnerSettings",
+]
